Use consistent placeholders in config.js
This changes GRAPHENDPOINT to an example value with a placeholder, while also replacing string literal values for the database and collection to be placeholders. This can help new users see at a glance that all configuration values should be set, whereas they might have left "graphdb" or "graphcoll" values alone.